PB-95: Moving utils components into a components subfolder
