+++ src/modules/infobox/components/FeatureElevationProfile.vue
@@ -41,8 +41,8 @@ import getProfile from '@/api/profile/profile.api'
 import { generateFilename } from '@/modules/drawing/lib/export-utils'
 import FeatureElevationProfileInformation from '@/modules/infobox/components/FeatureElevationProfileInformation.vue'
 import FeatureElevationProfilePlot from '@/modules/infobox/components/FeatureElevationProfilePlot.vue'
+import LoadingBar from '@/utils/components/LoadingBar.vue'
 import { LV95, WGS84 } from '@/utils/coordinates/coordinateSystems'
-import LoadingBar from '@/utils/LoadingBar.vue'
 import log from '@/utils/logging'
 
 export default {
