+++ src/modules/infobox/components/styling/DrawingStyleIconSelector.vue
@@ -79,7 +79,7 @@ import { useI18n } from 'vue-i18n'
 import { EditableFeature } from '@/api/features.api'
 import DrawingStyleColorSelector from '@/modules/infobox/components/styling/DrawingStyleColorSelector.vue'
 import DrawingStyleSizeSelector from '@/modules/infobox/components/styling/DrawingStyleSizeSelector.vue'
-import DropdownButton, { DropdownItem } from '@/utils/DropdownButton.vue'
+import DropdownButton, { DropdownItem } from '@/utils/components/DropdownButton.vue'
 import { MEDIUM } from '@/utils/featureStyleUtils'
 
 export default {
