+++ src/modules/map/components/footer/MapFooterAttributionList.vue
@@ -31,7 +31,7 @@
 import { mapGetters, mapState } from 'vuex'
 
 import MapFooterAttributionItem from '@/modules/map/components/footer/MapFooterAttributionItem.vue'
-import ThirdPartDisclaimer from '@/utils/ThirdPartDisclaimer.vue'
+import ThirdPartDisclaimer from '@/utils/components/ThirdPartDisclaimer.vue'
 
 export default {
     components: { MapFooterAttributionItem, ThirdPartDisclaimer },
