+++ src/modules/menu/components/header/HeaderWithSearch.vue
@@ -48,7 +48,7 @@ import HeaderSwissConfederationText from '@/modules/menu/components/header/Heade
 import SwissFlag from '@/modules/menu/components/header/SwissFlag.vue'
 import FeedbackToolbar from '@/modules/menu/components/menu/feedback/FeedbackToolbar.vue'
 import SearchBar from '@/modules/menu/components/search/SearchBar.vue'
-import LoadingBar from '@/utils/LoadingBar.vue'
+import LoadingBar from '@/utils/components/LoadingBar.vue'
 
 export default {
     components: {
