+++ src/modules/menu/components/menu/feedback/FeedbackButton.vue
@@ -106,8 +106,8 @@ import { mapGetters } from 'vuex'
 import sendFeedback from '@/api/feedback.api'
 import HeaderLink from '@/modules/menu/components/header/HeaderLink.vue'
 import FeedbackRating from '@/modules/menu/components/menu/feedback/FeedbackRating.vue'
+import ModalWithBackdrop from '@/utils/components/ModalWithBackdrop.vue'
 import log from '@/utils/logging'
-import ModalWithBackdrop from '@/utils/ModalWithBackdrop.vue'
 
 // comes from https://v2.vuejs.org/v2/cookbook/form-validation.html#Using-Custom-Validation
 const EMAIL_REGEX =
