+++ src/utils/components/OpenFullAppLink.vue
