+++ src/utils/components/ThirdPartDisclaimer.vue
@@ -20,7 +20,7 @@
 import tippy from 'tippy.js'
 import { mapState } from 'vuex'
 
-import ModalWithBackdrop from '@/utils/ModalWithBackdrop.vue'
+import ModalWithBackdrop from '@/utils/components/ModalWithBackdrop.vue'
 
 export default {
     components: { ModalWithBackdrop },
