+++ src/views/LoadingView.vue
@@ -9,7 +9,7 @@
 import { mapState } from 'vuex'
 
 import HeaderWithSearch from '@/modules/menu/components/header/HeaderWithSearch.vue'
-import LoadingBar from '@/utils/LoadingBar.vue'
+import LoadingBar from '@/utils/components/LoadingBar.vue'
 
 export default {
     components: { LoadingBar, HeaderWithSearch },
