+++ src/views/MapView.vue
@@ -22,8 +22,8 @@ import InfoboxModule from '@/modules/infobox/InfoboxModule.vue'
 import MapFooter from '@/modules/map/components/footer/MapFooter.vue'
 import MapModule from '@/modules/map/MapModule.vue'
 import MenuModule from '@/modules/menu/MenuModule.vue'
+import OpenFullAppLink from '@/utils/components/OpenFullAppLink.vue'
 import log from '@/utils/logging'
-import OpenFullAppLink from '@/utils/OpenFullAppLink.vue'
 
 export default {
     components: {
